import_task SnapshotDetails added
